Update for new Studierendenwerk website
The RSS feed is no longer available and there is a new web page showing
the menu for single days and for each location.

Each day is loaded consecutively for up to one week in advance. As the
menu may be changed, the app always tries to load today's menu if it
was last loaded more than 5 minutes ago. Thus, the Javascript code may
send a menu to the Pebble app multiple times.

Currently, only data for the Mensa in Vaihingen is loaded.